+++ inst/include/interface/rcpp/rcpp_objects/rcpp_population.hpp
@@ -130,7 +130,7 @@ class PopulationInterface : public PopulationInterfaceBase {
     population->maturity_id = this->maturity_id;
     population->log_M.resize(this->log_M.size());
     population->log_init_naa.resize(this->log_init_naa.size());
-    for (int i = 0; i < log_M.size(); i++) {
+    for (size_t i = 0; i < log_M.size(); i++) {
       population->log_M[i] = this->log_M[i].value_m;
       if (this->log_M[i].estimated_m) {
         info->RegisterParameter(population->log_M[i]);
@@ -138,7 +138,7 @@ class PopulationInterface : public PopulationInterfaceBase {
     }
     info->variable_map[this->log_M.id_m] = &(population)->log_M;
 
-    for (int i = 0; i < log_init_naa.size(); i++) {
+    for (size_t i = 0; i < log_init_naa.size(); i++) {
       population->log_init_naa[i] = this->log_init_naa[i].value_m;
       if (this->log_init_naa[i].estimated_m) {
         info->RegisterParameter(population->log_init_naa[i]);
